try circular deps
